truthmate-backend/routes/verify.py
from flask import Blueprint, request, jsonify
from models.verification import Verification
from models.user import User
from extensions import db
import jwt
import os
from services.fact_check import check_fact_with_google
from services.ai_detect import detect_ai_generated

verify_bp = Blueprint('verify', __name__)

@verify_bp.route('/text', methods=['POST'])
def verify_text():
    # Text verification logic
    pass

@verify_bp.route('/url', methods=['POST'])
def verify_url():
    # URL verification logic
    pass

@verify_bp.route('/image', methods=['POST'])
def verify_image():
    # Image verification logic
    pass