truthmate-backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
PyJWT==2.10.1
Werkzeug==2.3.7
requests==2.31.0
huggingface-hub==0.20.0